+++ backend/src/services/upload/nodeMigrator/index.ts
@@ -1,6 +1,6 @@
-import { logger } from '../../drivers/logger.js'
-import { UploadsUseCases } from '../../useCases/uploads/uploads.js'
-import { safeCallback } from '../../utils/safe.js'
+import { logger } from '../../../drivers/logger.js'
+import { UploadsUseCases } from '../../../useCases/uploads/uploads.js'
+import { safeCallback } from '../../../utils/safe.js'
 
 const state = {
   executing: false,
