+++ backend/src/services/upload/onchainPublisher/accounts.ts
@@ -1,5 +1,5 @@
 import { Keyring } from '@polkadot/api'
-import { config } from '../../config.js'
+import { config } from '../../../config.js'
 
 export const getAccounts = () => {
   const privateKeys = config.privateKeysPath
