+++ backend/src/worker.ts
@@ -1,6 +1,6 @@
-import { nodeMigrator } from './services/nodeMigrator/index.js'
-import { objectMappingArchiver } from './services/objectMappingListener/index.js'
-import { onchainPublisher } from './services/onchainPublisher/index.js'
+import { nodeMigrator } from './services/upload/nodeMigrator/index.js'
+import { objectMappingArchiver } from './services/dsn/objectMappingListener/index.js'
+import { onchainPublisher } from './services/upload/onchainPublisher/index.js'
 
 nodeMigrator.start()
 onchainPublisher.start()
